large-scale reorganization of code within the codebase
