+++ mmc_server.py
@@ -1,5 +1,5 @@
 import os
-from typing import Optional, Dict, List
+from typing import Dict, List
 import torch
 
 from fastapi import FastAPI
@@ -13,7 +13,7 @@ from tomcat_speech.data_prep.data_prep_helpers import (
 from tomcat_speech.train_and_test_models.asist_analysis_functions import predict_with_model
 
 # the current parameters file is saved as testing_parameters/config.py
-import tomcat_speech.train_and_test_models.testing_parameters.config as params
+import tomcat_speech.parameters.testing_parameters.config as params
 
 # Get model and glove paths
 MODEL_PATH = os.path.dirname(__file__) + "/data/MC_GOLD_classwts_nogender_25to75perc_avg_IS13.pth"
