+++ tomcat_speech/evaluation_scripts/evaluate_unlabeled_data.py
@@ -1,29 +1,26 @@
-# test a single-task model for any of the five tasks
-# combining the testing of the separate train_task models into one
-
 # train the models created in models directory with MUStARD data
 # currently the main entry point into the system
-import pickle
 import shutil
 import sys
+sys.path.append("/")
 import os
+
+import pandas as pd
 import torch
 import numpy as np
-from datetime import date
 import random
 
-from tomcat_speech.models.train_and_test_models import evaluate
 from tomcat_speech.train_and_test_models.train_and_test_utils import make_train_state
+from tomcat_speech.train_and_test_models.train_and_test_without_gold_labels import multitask_predict_without_gold_labels
 from tomcat_speech.models.multimodal_models import MultitaskModel
 
 # import MultitaskObject and Glove from preprocessing code
-sys.path.append("../multimodal_data_preprocessing")
+sys.path.append("/home/jculnan/github/multimodal_data_preprocessing")
 from utils.data_prep_helpers import (
-    MultitaskObject,
     Glove,
     make_glove_dict,
-    MultitaskTestObject,
 )
+from tomcat_speech.training_scripts.train_multitask import load_modality_data
 
 # set device
 cuda = False
@@ -35,6 +32,7 @@ device = torch.device("cuda" if cuda else "cpu")
 # # Check CUDA
 if torch.cuda.is_available():
     torch.cuda.set_device(2)
+
 
 if __name__ == "__main__":
     # set saved model
@@ -48,10 +46,12 @@ if __name__ == "__main__":
     config_path = os.path.abspath(config_path)
     output_path = os.path.dirname(config_path)
 
-    shutil.copyfile(config_path, "mmml/config_files/test_config.py")
+    shutil.copyfile(
+        config_path, "tomcat_speech/train_and_test_models/testing_parameters/config.py"
+    )
 
     # import parameters for model
-    import mmml.config_files.test_config as config
+    from tomcat_speech.parameters.testing_parameters import config
 
     # set random seed
     torch.manual_seed(config.model_params.seed)
@@ -69,7 +69,7 @@ if __name__ == "__main__":
 
     # decide if you want to use avgd feats
     avgd_acoustic_in_network = (
-        config.model_params.avgd_acoustic or config.model_params.add_avging
+            config.model_params.avgd_acoustic or config.model_params.add_avging
     )
 
     # set location for pickled data (saving or loading)
@@ -81,19 +81,31 @@ if __name__ == "__main__":
     # add stdout to a log file
     with open(os.path.join(output_path, "log"), "a") as f:
         if not config.DEBUG:
-            sys.stdout = f
+            # sys.stdout = f
+
+            # separate from training/dev code
+            print("\n")
+            print("BEGINNING EVALUATION OF TRAINED MODEL")
 
             # 1. Load datasets + glove object
             load_path = config.load_path
             feature_set = config.feature_set
 
-            # get
-            test_ds = pickle.load(
-                open(f"{load_path}/{config.task}_{feature_set}_test.pickle", "rb")
-            )
-            class_weights = pickle.load(
-                open(f"{load_path}/{config.task}_{feature_set}_clsswts.pickle", "rb")
-            )
+            # import data
+            #used_ds = pickle.load(open(f"{load_path}", "rb"))
+            #print("Data loaded")
+
+            # if data is field-separated
+            used_ds = load_modality_data(device, config,
+                                         use_text=True,
+                                         use_acoustic=True)
+            #used_ds = used_ds[0][0].train
+            all_ds = []
+            all_ds.extend(used_ds[0][0].train)
+            all_ds.extend(used_ds[0][0].dev)
+            all_ds.extend(used_ds[0][0].test)
+            used_ds = all_ds
+            print("Data loaded")
 
             # if not using distilbert embeddings
             if not config.model_params.use_distilbert:
@@ -106,42 +118,33 @@ if __name__ == "__main__":
                 num_embeddings = pretrained_embeddings.size()[0]
                 print(f"shape of pretrained embeddings is: {glove.data.size()}")
 
-            # 3. CREATE NN
             # this uses train-dev-test folds
             # create instance of model
             if config.model_params.use_distilbert:
-                task_model = MultitaskModel(
+                multitask_model = MultitaskModel(
                     params=config.model_params,
                     use_distilbert=config.model_params.use_distilbert,
+                    multi_dataset=False
                 )
             else:
-                task_model = MultitaskModel(
+                multitask_model = MultitaskModel(
                     params=config.model_params,
                     use_distilbert=config.model_params.use_distilbert,
                     num_embeddings=num_embeddings,
                     pretrained_embeddings=pretrained_embeddings,
+                    multi_dataset=False
                 )
 
             optimizer = torch.optim.Adam(
                 lr=config.model_params.lr,
-                params=task_model.parameters(),
+                params=multitask_model.parameters(),
                 weight_decay=config.model_params.weight_decay,
             )
 
             # set the classifier(s) to the right device
             # get saved parameters
-            task_model.load_state_dict(torch.load(saved_model, map_location=device))
-            task_model.to(device)
-
-            # add loss function for cdc
-            loss_func = torch.nn.CrossEntropyLoss(
-                weight=class_weights.to(device), reduction="mean"
-            )
-
-            task_obj = MultitaskTestObject(test_ds, loss_func, task_num=0)
-
-            # set all data list
-            all_data_list = [task_obj]
+            multitask_model.load_state_dict(torch.load(saved_model, map_location=device))
+            multitask_model.to(device)
 
             print("Trained model loaded, loss function, and optimization prepared")
 
@@ -155,15 +158,36 @@ if __name__ == "__main__":
                 config.model_params.early_stopping_criterion,
             )
 
-            # train the model and evaluate on development set
-            evaluate(
-                task_model,
-                train_state,
-                all_data_list,
+            # get order of predictions
+            #ordered_ids = [item['audio_id'] for item in used_ds[0]]
+
+            # get predictions
+            ordered_predictions, ordered_penult_lyrs, ordered_ids = multitask_predict_without_gold_labels(
+                multitask_model,
+                used_ds,
                 config.model_params.batch_size,
-                pickle_save_name=f"{item_output_path}/test_results.pickle",
-                device=device,
+                device,
+                num_predictions=3,
                 avgd_acoustic=avgd_acoustic_in_network,
                 use_speaker=config.model_params.use_speaker,
                 use_gender=config.model_params.use_gender,
+                get_prob_dist=True,
+                return_penultimate_layer=True,
+                use_spec=config.model_params.use_spec,
             )
+
+            all_preds = {}
+            for pred_type in ordered_predictions.keys():
+                all_preds[pred_type] = []
+                #print(pred_type)
+                #print("--------------------")
+                for group_pred in ordered_predictions[pred_type]:
+                    for preds in group_pred:
+                        all_preds[pred_type].append(preds.index(max(preds)))
+                        print(preds.index(max(preds)))
+
+            data_to_save = pd.DataFrame(all_preds)
+            data_to_save.columns = ['trait', 'emotion', 'sentiment']
+            data_to_save['audio_id'] = ordered_ids
+            #data_to_save.to_csv(f"../../PROJECTS/ToMCAT/Evaluating_modelpredictions/data_from_speechAnalyzer/used_for_evaluating_model_results/all_preds_no_classweights.csv", index=False)
+            data_to_save.to_csv(f"{config.exp_save_path}/_predictions.csv", index=False)
