+++ tomcat_speech/train_and_test_models/train_and_test_personality_multitask.py
@@ -1,4 +1,11 @@
 
+from torch.utils.data import DataLoader
+from sklearn.metrics import confusion_matrix
+from sklearn.metrics import classification_report
+from sklearn.metrics import precision_recall_fscore_support
+
+from tomcat_speech.train_and_test_models.train_and_test_utils import update_train_state
+
 
 def personality_as_multitask_train_and_predict(
     classifier,
